gallery
I'm almost ashamed to propose this but still, it's still something. It's pretty much my first html/css project and definitely my first js function, so please be indulgent ! It's a quick put together if you want a gallery view for pictures of events.

There's a gallery on the left hand side that can be scrolled (pure CSS) and a central view for the image displayed through "mouseover" action above the thumbnails. (I used the links for the first few images of the Facebook Hackathon, just to test it out. Probably an automated js method to automate the whole thing but idk how). The script is just to replicate the color change like you have on your website between the Hackers and @EPFL, thought it was cool. 
I did not add captions since i did not know what to write in them ...

Please note once again this is far from perfect (and has absolutely no prevention to be even close to that), and there is plenty to correct, but i'm only learning, and i'd be more than glad to hear critique and advice.